Refactor into different files JC-622
