+++ apps/vanilla/src/pages/blackjack/PlayerTurnResult.ts
@@ -1,0 +1,5 @@
+/** This is foo comment . */
+export class PlayerTurnResult {
+	/** This is foo comment . */
+	public constructor(public playerIndex: number, public diceResult: number) {}
+}
