+++ apps/vanilla/src/pages/blackjack/index.ts
@@ -2,135 +2,11 @@
 /* eslint-disable jsdoc/require-jsdoc */
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
 /* eslint-disable no-console */
-import { Subscriber } from './Subscriber';
-
-/** Publisher class . */
-class Publisher<T> {
-	/** Array of subscribers . */
-	private subscribers: Subscriber<T>[] = [];
-
-	/** Method to add subscriber . */
-	public subscribe(subscriber: Subscriber<T>): void {
-		this.subscribers.push(subscriber);
-	}
-
-	/** Method to remove subscriber . */
-	public unsubscribe(subscriber: Subscriber<T>): void {
-		this.subscribers = this.subscribers.filter((sub) => sub !== subscriber);
-	}
-
-	/** Method to notify . */
-	public notify(message: T): void {
-		this.subscribers.forEach((sub) => sub.update(message));
-	}
-}
-
-/** This is foo comment . */
-class TurnGenerator extends Publisher<number> {
-	private playerCount;
-
-	public currentPlayerIndex = 0;
-
-	public constructor(private plCount: number) {
-		super();
-		this.playerCount = this.plCount;
-	}
-
-	/** Next . */
-	public next(): void {
-		const playerIndex = this.currentPlayerIndex;
-		this.currentPlayerIndex = (playerIndex + 1) % this.playerCount;
-		this.notify();
-	}
-
-	public override notify(): void {
-		super.notify(this.currentPlayerIndex);
-	}
-}
-
-/** This is foo comment . */
-class DiceGenerator extends Publisher<PlayerTurnResult> implements Subscriber<number> {
-	/** */
-	public sideCount = 6;
-
-	public roll(): number {
-		return Math.floor(Math.random() * this.sideCount) + 1;
-	}
-
-	public update(playerIndex: number) {
-		const diceResult = this.roll();
-		const turnResult = new PlayerTurnResult(playerIndex, diceResult);
-		this.notify(turnResult);
-	}
-}
-
-/** This is foo comment . */
-class PlayerTurnResult {
-	/** This is foo comment . */
-	public constructor(public playerIndex: number, public diceResult: number) {}
-}
-
-/** This is foo comment . */
-class Player implements Subscriber<PlayerTurnResult> {
-	private readonly diceResults: number[] = [];
-
-	public results: Publisher<number[]> = new Publisher<number[]>();
-
-	public winStatus: Publisher<boolean> = new Publisher<boolean>();
-
-	public constructor(private playerIndex: number) {}
-
-	private getTotalScore(): number {
-		return this.diceResults.reduce((sum, result) => sum + result, 0);
-	}
-
-	/** This is foo comment . */
-	public update(message: PlayerTurnResult): void {
-		if (message.playerIndex === this.playerIndex) {
-			this.diceResults.push(message.diceResult);
-			this.results.notify(this.diceResults);
-			if (this.getTotalScore() > 21) {
-				this.winStatus.notify(true);
-			}
-		}
-	}
-}
-
-/** This is foo comment . */
-class ResultDisplayer implements Subscriber<number[]> {
-	/** This is foo comment . */
-	private diceResult: number[] = [];
-
-	private totalScores = 0;
-
-	private getTotalScore(): number {
-		return this.diceResult.reduce((sum, result) => sum + result, 0);
-	}
-
-	public constructor(private resultElement: HTMLElement | null, private totalElement: HTMLElement | null) {}
-
-	public update(message: number[]): void {
-		this.diceResult = message;
-		this.totalScores = this.getTotalScore();
-		if (this.resultElement) {
-			this.resultElement.innerText = `${this.diceResult.join(', ')}`;
-		}
-		if (this.totalElement) {
-			this.totalElement.innerText = `${this.totalScores}`;
-		}
-	}
-}
-
-class WinnerDisplayer implements Subscriber<boolean> {
-	public constructor(private winnerElement: HTMLElement | null) {}
-
-	public update(message: boolean): void {
-		console.log(message);
-		if (message && this.winnerElement) {
-			this.winnerElement.classList.add('won');
-		}
-	}
-}
+import { DiceGenerator } from './DiceGenerator';
+import { Player } from './Player';
+import { ResultDisplayer } from './ResultDisplayer';
+import { TurnGenerator } from './TurnGenerator';
+import { WinnerDisplayer } from './WinnerDisplayer';
 
 const player1DiceResult = document.getElementById('0');
 const player2DiceResult = document.getElementById('1');
